images are hotel part; back browser button works; home button added
