+++ admin/src/app/Components/change-password/change-password.component.html
@@ -63,7 +63,6 @@
             >
                 Submit
             </button>
-            <button mat-button routerLink='/hotels'>Back</button>
         </mat-card-actions>
     </mat-card>
 </form>
