+++ admin/src/app/Components/navbar/navbar.component.css
@@ -3,3 +3,8 @@ nav {
     margin-right: 15%;
     padding: 2%
 }
+
+button {
+    background-color: #B814D8;
+    color: white;
+}
