+++ admin/src/app/Core/pageParameters.ts
@@ -1,4 +1,6 @@
 import { HttpParams } from '@angular/common/http';
+import { PageEvent } from '@angular/material/paginator';
+import { Params } from '@angular/router';
 
 export class PageParameters {
   public constructor(
@@ -8,17 +10,24 @@ export class PageParameters {
   }
 
   public getHttpParams(): HttpParams{
-    let params = new HttpParams()
+    const params = new HttpParams()
       .set('pageIndex', this.pageIndex)
       .set('pageSize', this.pageSize);
     return params;
   }
 
   public getHttpParamsObj(): object{
-    let params = {
+    const params = {
       'pageIndex': this.pageIndex,
       'pageSize': this.pageSize
     };
     return params;
+  }
+
+  public updateParameters(data: PageEvent | Params): void {
+    if (data.pageIndex && data.pageSize) {
+      this.pageIndex = data.pageIndex;
+      this.pageSize = data.pageSize;
+    }
   }
 }
